REFACTOR : 검색 창 따로 분리
